chat service - serviceImpl

+++ src/main/java/com/sports/Chat/ChatRoomRepository.java
@@ -2,6 +2,6 @@ package com.sports.Chat;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChatRoomRepository extends JpaRepository<ChatRoom, Integer> {
+public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
 
 }
